make more changes

+++ src/main/java/Services/CustomerService.java
@@ -5,27 +5,53 @@ import Entity.*;
 
 
 public class CustomerService extends EntityService<Customer> {
-    private final PermissionService permission;
-    private final ProductService productService;
+    private PermissionService permission;
+    private ProductService productService;
+    private OrderService orderService;
 
-    public CustomerService(AuthService authService, PermissionService permission, ProductService productService) {
+    public CustomerService(AuthService authService, PermissionService permission, ProductService productService, OrderService orderService) {
         super("customers", Customer.class, authService);
         this.permission = permission;
         this.productService = productService;
+        this.orderService = orderService;
     }
 
-    public void create(String username, String password, java.util.Date dateOfBirth, double balance, String address, Gender gender) {        
-        if (permission.hasPermission("customers", "create")) {
+    public PermissionService getPermission() {
+        return permission;
+    }
+
+    public void setPermission(PermissionService permission) {
+        this.permission = permission;
+    }
+
+    public ProductService getProductService() {
+        return productService;
+    }
+
+    public void setProductService(ProductService productService) {
+        this.productService = productService;
+    }
+
+    public OrderService getOrderService() {
+        return orderService;
+    }
+
+    public void setOrderService(OrderService orderService) {
+        this.orderService = orderService;
+    }
+
+    public String create(String username, String password, java.util.Date dateOfBirth, double balance, String address, Gender gender) {        
+        if (permission.hasPermission("customers", "create")||getLoggedInUser()==null) {
             if (getEntityDAO().getIndex(username)!=-1) {
                 throw new IllegalArgumentException("This username is used");
             }
 
-            EntityDAO<Order> orderDAO = new EntityDAO<>("orders", Order.class);
-            String cartId = orderDAO.nextId();
-            Order cart = new Order(cartId, username, new HashMap<>(), balance, null, Status.draft);
-            Customer user = new Customer(username,password,dateOfBirth,balance,address,gender,new ArrayList<>(),cartId);    
+            Customer user = new Customer(username,password,dateOfBirth,balance,address,gender,new ArrayList<>());    
             getEntityDAO().add(user);
-            orderDAO.add(cart);
+            String cartId = orderService.create(username, new HashMap<>(), null, Status.draft);
+            user.setCartId(cartId);
+            getEntityDAO().update(user);
+            return username;
         }else{
             throw new RuntimeException("You don't have the permisson to do this action");
         }
@@ -34,17 +60,31 @@ public class CustomerService extends EntityService<Customer> {
     public void delete(String username) {
         if (permission.hasPermission("customers", "delete")) {
             Customer customer = getEntityDAO().get(username);
-            EntityDAO<Order> orderDAO = new EntityDAO<>("orders", Order.class);
+            if (customer == null) {
+                throw new IllegalArgumentException("This customer doesn't exist.");  
+            }
             this.getEntityDAO().delete(username);
-            orderDAO.delete(customer.getCartId());
+            orderService.delete(customer.getCartId());
         }else{
             throw new RuntimeException("You don't have the permisson to do this action");
         }
     }
 
     public Customer get(String username){
+        if (getLoggedInUser() == null) {
+            Customer customer = getEntityDAO().get(username);
+            if (customer == null) {
+                return null;
+            }
+            return new Customer(customer);
+        }
+        
         if (permission.hasPermission("customers","retrieve")|| getLoggedInUser().equals(getEntityDAO().get(username))) {
-            return new Customer(getEntityDAO().get(username));
+            Customer customer = getEntityDAO().get(username);
+            if (customer == null) {
+                return null;
+            }
+            return new Customer(customer);
         }else{
             throw new RuntimeException("You don't have the permisson to do this action");
         }
@@ -61,6 +101,10 @@ public class CustomerService extends EntityService<Customer> {
     public <T> void update(String username,String parameter ,T newData) {
         if (permission.hasPermission("customers", "update")||getLoggedInUser().equals(getEntityDAO().get(username))) {
             Customer customer = getEntityDAO().get(username);
+            if (customer == null) {
+                throw new IllegalArgumentException("This customer doesn't exist.");  
+            }
+
             switch (parameter.toLowerCase()) {
                 case "password":
                     customer.setPassword((String)newData);
@@ -84,10 +128,11 @@ public class CustomerService extends EntityService<Customer> {
         if (permission.hasPermission("customers", "update")||getLoggedInUser().equals(getEntityDAO().get(username))){
             productService.get(productId);
             Customer customer = getEntityDAO().get(username);
-            EntityDAO<Order> orderDAO = new EntityDAO<>("orders", Order.class);
-            Order cart = orderDAO.get(customer.getCartId());
-            cart.addProduct(productId);
-            orderDAO.update(cart);
+            if (customer == null) {
+                throw new IllegalArgumentException("This customer doesn't exist.");  
+            }
+
+            orderService.addItem(customer.getCartId(), productId);
         }else{
             throw new RuntimeException("You don't have the permisson to do this action");
         }
@@ -97,10 +142,10 @@ public class CustomerService extends EntityService<Customer> {
         if (permission.hasPermission("customers", "update")||getLoggedInUser().equals(getEntityDAO().get(username))){
             productService.get(productId);
             Customer customer = getEntityDAO().get(username);
-            EntityDAO<Order> orderDAO = new EntityDAO<>("orders", Order.class);
-            Order cart = orderDAO.get(customer.getCartId());
-            cart.removeProduct(productId);
-            orderDAO.update(cart);
+            if (customer == null) {
+                throw new IllegalArgumentException("This customer doesn't exist.");  
+            }
+            orderService.deleteItem(customer.getCartId(), productId);
         }else{
             throw new RuntimeException("You don't have the permisson to do this action");
         }
@@ -109,9 +154,10 @@ public class CustomerService extends EntityService<Customer> {
     public Map<String, Integer> getCartProducts(String username){
         if (permission.hasPermission("customers", "update")||getLoggedInUser().equals(getEntityDAO().get(username))){
             Customer customer = getEntityDAO().get(username);
-            EntityDAO<Order> orderDAO = new EntityDAO<>("orders", Order.class);
-            Order cart = orderDAO.get(customer.getCartId());
-            return cart.getProducts();
+            if (customer == null) {
+                throw new IllegalArgumentException("This customer doesn't exist.");  
+            }
+            return orderService.get(customer.getCartId()).getProducts();
         }else{
             throw new RuntimeException("You don't have the permisson to do this action");
         }
@@ -121,6 +167,9 @@ public class CustomerService extends EntityService<Customer> {
         if (permission.hasPermission("customers", "update")||getLoggedInUser().equals(getEntityDAO().get(username))){
             productService.get(productId);
             Customer customer = getEntityDAO().get(username);
+            if (customer == null) {
+                throw new IllegalArgumentException("This customer doesn't exist.");  
+            }
             ArrayList<String> products = customer.getInterests();
             if (products.contains(productId)) {
                 throw new RuntimeException("This product is already in interests");
@@ -137,6 +186,9 @@ public class CustomerService extends EntityService<Customer> {
         if (permission.hasPermission("customers", "update")||getLoggedInUser().equals(getEntityDAO().get(username))){
             productService.get(productId);
             Customer customer = getEntityDAO().get(username);
+            if (customer == null) {
+                throw new IllegalArgumentException("This customer doesn't exist.");  
+            }
             ArrayList<String> products = customer.getInterests();
             products.remove(productId);
             customer.setInterests(products);
